pyproject.toml
[project]
name = "atlas_stream"
version = "0.3.0"
authors = [{ name = "dbbruijn@gmail.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    # Runtime dependencies — included in the deployed wheel and job environments.
    # See also https://docs.databricks.com/dev-tools/bundles/library-dependencies
    "requests>=2.31,<3",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "responses",
    "ruff",
    "databricks-connect>=15.4,<15.5",
    "ipykernel",
]

[project.scripts]
# ETL pipeline entry points — used by python_wheel_task in the job
ingest_countries          = "atlas_stream.etl.ingest:main"
bronze_countries_to_silver = "atlas_stream.etl.bronze_to_silver:main"
silver_countries_to_gold  = "atlas_stream.etl.silver_to_gold:main"
verify_country_tables     = "atlas_stream.etl.verify:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "spark: marks tests that require a live Databricks Connect session (deselect with -m 'not spark')",
]
addopts = "--cov=src/atlas_stream --cov-report=term-missing"

[tool.coverage.run]
omit = ["src/**/*.ipynb"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
# F821: notebooks use Databricks runtime-injected globals (spark, dbutils)
per-file-ignores = { "src/**/*.ipynb" = ["F821"] }
